requirements.txt
anthropic
python-dotenv
requests
streamlit